clean up board coordinates

+++ include/dctl/evaluate/feature.hpp
@@ -39,7 +39,7 @@ public:
         {
                 using namespace xstd::support_literals;
                 int score = 0;
-                for (auto i = 1_z; i < static_cast<std::size_t>(Board::height()); ++i)
+                for (auto i = 1_z; i < Board::height(); ++i)
                         score += Weight<Rules, Board>::tempo[i] * static_cast<int>((p.pieces(ToMove) & board::Row<Board>::mask(ToMove, i)).count());
                 return score;
         }
@@ -49,7 +49,7 @@ public:
         {
                 using namespace xstd::support_literals;
                 int score = 0;
-                for (auto i = 1_z; i < static_cast<std::size_t>(Board::width()) / 2; ++i) {
+                for (auto i = 1_z; i < Board::width() / 2; ++i) {
                         score += Weight<Rules, Board>::center[i] *
                         (
                                 static_cast<int>((p.pieces(ToMove) & board::Column<Board>::mask( ToMove, i)).count()) +
@@ -64,7 +64,7 @@ public:
         {
                 using namespace xstd::support_literals;
                 int score = 0;
-                for (auto i = 0_z; i < static_cast<std::size_t>(Board::width()) / 2; ++i) {
+                for (auto i = 0_z; i < Board::width() / 2; ++i) {
                         score += Weight<Rules, Board>::balance[i] *
                         (
                                 static_cast<int>((p.pieces(ToMove) & board::Column<Board>::mask( ToMove, i)).count()) -
